+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/api/storage/INetworked.java
@@ -18,4 +18,6 @@ public interface INetworked
     }
 
     default void unbindControlGroup () { }
+
+    default void scheduleValidation () { }
 }