+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/tiledata/UpgradeData.java
@@ -1,6 +1,7 @@
 package com.jaquadro.minecraft.storagedrawers.block.tile.tiledata;
 
 import com.jaquadro.minecraft.storagedrawers.api.storage.IDrawerAttributesModifiable;
+import com.jaquadro.minecraft.storagedrawers.api.storage.attribute.MagnetDim;
 import com.jaquadro.minecraft.storagedrawers.config.ModCommonConfig;
 import com.jaquadro.minecraft.storagedrawers.core.ModItems;
 import com.jaquadro.minecraft.storagedrawers.item.*;
@@ -19,6 +20,7 @@ public class UpgradeData extends BlockEntityDataShim
 {
     protected final ItemStack[] upgrades;
     private int storageMultiplier;
+    private int[] magnetRange;
     private EnumUpgradeRedstone redstoneType;
 
     // TODO: Do we need to provide these?
@@ -30,6 +32,8 @@ public class UpgradeData extends BlockEntityDataShim
     private boolean hasIllumination;
     private boolean hasFillLevel;
     private boolean hasBalanceFill;
+    private boolean hasHopper;
+    private boolean hasMagnet;
     private boolean hasRemote;
 
     private IDrawerAttributesModifiable attrs;
@@ -39,6 +43,7 @@ public class UpgradeData extends BlockEntityDataShim
         Arrays.fill(upgrades, ItemStack.EMPTY);
 
         syncStorageMultiplier();
+        syncMagnetRange();
     }
 
     public void setDrawerAttributes (IDrawerAttributesModifiable attrs) {
@@ -152,6 +157,14 @@ public class UpgradeData extends BlockEntityDataShim
         return storageMultiplier;
     }
 
+    public int getMagnetRange (MagnetDim dim) {
+        return switch (dim) {
+            case MagnetDim.HORIZONTAL -> magnetRange[0];
+            case MagnetDim.UP -> magnetRange[1];
+            case MagnetDim.DOWN -> magnetRange[2];
+        };
+    }
+
     public EnumUpgradeRedstone getRedstoneType () {
         return redstoneType;
     }
@@ -179,6 +192,10 @@ public class UpgradeData extends BlockEntityDataShim
     public boolean hasbalancedFillUpgrade () {
         return hasBalanceFill;
     }
+
+    public boolean hasHopperUpgrade () { return hasHopper; }
+
+    public boolean hasMagnetUpgrade () { return hasMagnet; }
 
     public boolean hasRemoteUpgrade () {
         return hasRemote;
@@ -240,6 +257,7 @@ public class UpgradeData extends BlockEntityDataShim
 
         syncStorageMultiplier();
         syncRedstoneLevel();
+        syncMagnetRange();
 
         hasOneStack = false;
         hasVoid = false;
@@ -249,6 +267,7 @@ public class UpgradeData extends BlockEntityDataShim
         hasIllumination = false;
         hasFillLevel = false;
         hasBalanceFill = false;
+        hasHopper = false;
         hasRemote = false;
 
         for (ItemStack stack : upgrades) {
@@ -270,6 +289,8 @@ public class UpgradeData extends BlockEntityDataShim
                 hasFillLevel = ModCommonConfig.INSTANCE.UPGRADES.fillLevelUpgrade.enableUpgrade.get();
             else if (item == ModItems.BALANCE_FILL_UPGRADE.get())
                 hasBalanceFill = ModCommonConfig.INSTANCE.UPGRADES.balanceUpgrade.enableUpgrade.get();
+            else if (item == ModItems.HOPPER_UPGRADE.get())
+                hasHopper = ModCommonConfig.INSTANCE.UPGRADES.balanceUpgrade.enableUpgrade.get();
             else if (item instanceof ItemUpgradeRemote remote) {
                 boolean enable = ModCommonConfig.INSTANCE.UPGRADES.remoteUpgrade.enableUpgrade.get();
                 hasRemote = remote.isGroupUpgrade()
@@ -284,6 +305,8 @@ public class UpgradeData extends BlockEntityDataShim
         attrs.setIsUnlimitedStorage(hasUnlimited);
         attrs.setIsUnlimitedVending(hasVending);
         attrs.setIsBalancedFill(hasBalanceFill);
+        attrs.setIsHopper(hasHopper);
+        attrs.setIsMagnet(hasMagnet);
     }
 
     private void syncStorageMultiplier () {
@@ -298,6 +321,26 @@ public class UpgradeData extends BlockEntityDataShim
 
         if (storageMultiplier == 0)
             storageMultiplier = ModCommonConfig.INSTANCE.UPGRADES.getLevelMult(0);
+    }
+
+    private void syncMagnetRange () {
+        if (magnetRange == null || magnetRange.length != 3)
+            magnetRange = new int[3];
+
+        Arrays.fill(magnetRange, 0);
+        hasMagnet = false;
+
+        for (ItemStack stack : upgrades) {
+            if (stack.getItem() instanceof ItemUpgradeMagnet itemMagnet) {
+                if (!itemMagnet.isEnabled())
+                    continue;
+
+                hasMagnet = true;
+                magnetRange[0] += itemMagnet.getHorzRange();
+                magnetRange[1] += itemMagnet.getUpRange();
+                magnetRange[2] += itemMagnet.getDownRange();
+            }
+        }
     }
 
     private void syncRedstoneLevel () {